Fixed all Elemento logic using DTO's when needed and related services and controllers, implemented full carrello and ordine logic

+++ FilieraAgricola/src/main/java/it/unicam/cs/filieraagricola/api/controller/elemento/ProdottoBaseController.java
@@ -1,6 +1,4 @@
 package it.unicam.cs.filieraagricola.api.controller.elemento;
-
-import it.unicam.cs.filieraagricola.api.entities.Users;
 import it.unicam.cs.filieraagricola.api.entities.elemento.Categoria;
 import it.unicam.cs.filieraagricola.api.entities.elemento.ProdottoBase;
 import it.unicam.cs.filieraagricola.api.services.UserService;
@@ -11,20 +9,18 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import static it.unicam.cs.filieraagricola.api.commons.ResponseEntityUtil.unauthorizedResponse;
+
 @RestController
 @RequestMapping("/operatore/produttore")
 public class ProdottoBaseController {
+
     @Autowired
     private ProdottoBaseService prodottoBaseService;
     @Autowired
     private UserService userService;
     @Autowired
     private CategoriaService categoriaService;
-
-    @GetMapping("/dashboard")
-    public ResponseEntity<String> getDashboard() {
-        return ResponseEntity.ok("Benvenuto nella dashboard del Produttore");
-    }
 
     @RequestMapping({"/prodottibase"})
     public ResponseEntity<Object> getProdottiBase() {
@@ -35,19 +31,20 @@ public class ProdottoBaseController {
     }
 
     @RequestMapping({"/prodottibase/{id}"})
-    public ResponseEntity<Object> getProdottoBase(@PathVariable("id") int id) {
+    public ResponseEntity<Object> getProdottoBase(@PathVariable int id) {
         if(prodottoBaseService.getProdottoBase(id) != null){
             if(prodottoBaseService.getProdottoBase(id).getOperatore() == userService.getCurrentUser()) {
                 return new ResponseEntity<>(prodottoBaseService.getProdottoBase(id), HttpStatus.FOUND);
-            } else {
-                return new ResponseEntity<>("Non sei autorizzato a vedere questo Prodotto", HttpStatus.UNAUTHORIZED);
-            }
+            } else return unauthorizedResponse();
         }
         return new ResponseEntity<>("Prodotto base non esistente", HttpStatus.NOT_FOUND);
     }
 
     @PostMapping({"/prodottibase/aggiungi"})
     public ResponseEntity<Object> aggiungiProdottoBase(@RequestBody ProdottoBase prodotto) {
+        if(!categoriaService.existsSameCategoria(prodotto.getCategoria().getNome())){
+            return new ResponseEntity<>("Categoria non esistente, Categorie esistenti: " + categoriaService.getAllCategorie(), HttpStatus.NOT_FOUND);
+        }
         if(prodottoBaseService.aggiungiProdottoBase(prodotto)){
             return new ResponseEntity<>("Prodotto creato", HttpStatus.CREATED);
         }
@@ -72,26 +69,60 @@ public class ProdottoBaseController {
         return new ResponseEntity<>("Prodotto già esistente", HttpStatus.CONFLICT);
     }
 
-    @DeleteMapping({"/prodottibase/elimina/{id}"})
-    public ResponseEntity<Object> deleteProduct(@PathVariable("id") int id) {
-        if(prodottoBaseService.getProdottoBase(id).getOperatore() != userService.getCurrentUser()) {
-            return new ResponseEntity<>("Non sei autorizzato", HttpStatus.UNAUTHORIZED);
-        }
-        prodottoBaseService.deleteProdottoBase(id);
-        return new ResponseEntity<>("Prodotto eliminato", HttpStatus.OK);
+    @DeleteMapping({"/prodottibase/elimina"})
+    public ResponseEntity<Object> eliminaProdotto(@RequestParam int id) {
+        if(prodottoBaseService.getProdottoBase(id) != null) {
+            if (prodottoBaseService.getProdottoBase(id).getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            prodottoBaseService.deleteProdottoBase(id);
+            return new ResponseEntity<>("Prodotto eliminato", HttpStatus.OK);
+        } else return new ResponseEntity<>("Prodotto non esistente", HttpStatus.NOT_FOUND);
     }
 
-    @RequestMapping(
-            value = {"/prodottibase/aggiorna"},
-            method = {RequestMethod.PUT}
-    )
+    @PutMapping("/prodottibase/aggiorna")
     public ResponseEntity<Object> aggiornaProdottoBase(@RequestBody ProdottoBase prodottoBase) {
-        if(prodottoBase.getOperatore() != userService.getCurrentUser()) {
-            return new ResponseEntity<>("Non sei autorizzato", HttpStatus.UNAUTHORIZED);
-        }
-        if(prodottoBaseService.aggiornaProdottoBase(prodottoBase)){
-            return new ResponseEntity<>("Prodotto aggiornato", HttpStatus.FOUND);
+        if(prodottoBaseService.existsProdottoBase(prodottoBase.getId())){
+            ProdottoBase prodotto = prodottoBaseService.getProdottoBase(prodottoBase.getId());
+            if (prodotto.getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            if (prodottoBaseService.aggiornaProdottoBase(prodottoBase)) {
+                return new ResponseEntity<>("Prodotto aggiornato", HttpStatus.FOUND);
+            }
+            return new ResponseEntity<>("Prodotto non aggiornato", HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>("Prodotto non esistente", HttpStatus.NOT_FOUND);
+    }
+
+    @PutMapping("/prodottibase/aggiorna-quantita")
+    public ResponseEntity<Object> aggiornaQuantitaProdotto(@RequestParam int id, @RequestParam int quantita) {
+        ProdottoBase prodotto = prodottoBaseService.getProdottoBase(id);
+        if (prodotto != null) {
+            if (prodotto.getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            if (prodottoBaseService.aggiornaQuantitaProdotto(id, quantita)) {
+                return new ResponseEntity<>("Quantità aggiornata", HttpStatus.OK);
+            }
+            return new ResponseEntity<>("Impossibile aggiornare la quantità", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>("Prodotto non trovato", HttpStatus.NOT_FOUND);
+    }
+
+    @PutMapping("/prodottibase/rimuovi-quantita")
+    public ResponseEntity<Object> rimuoviQuantitaProdotto(@RequestParam int id, @RequestParam int quantita) {
+        ProdottoBase prodotto = prodottoBaseService.getProdottoBase(id);
+        if (prodotto != null) {
+            if (prodotto.getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            if (prodottoBaseService.riduciQuantitaProdotto(id, quantita)) {
+                return new ResponseEntity<>("Quantità ridotta", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Quantità insufficiente per rimuovere", HttpStatus.BAD_REQUEST);
+            }
+        }
+        return new ResponseEntity<>("Prodotto non trovato", HttpStatus.NOT_FOUND);
     }
 }
