+++ FilieraAgricola/src/main/java/it/unicam/cs/filieraagricola/api/controller/elemento/ProdottoTrasformatoController.java
@@ -1,8 +1,12 @@
 package it.unicam.cs.filieraagricola.api.controller.elemento;
 
+import it.unicam.cs.filieraagricola.api.commons.richiesta.StatoContenuto;
 import it.unicam.cs.filieraagricola.api.entities.elemento.Categoria;
+import it.unicam.cs.filieraagricola.api.entities.elemento.ProdottoBase;
 import it.unicam.cs.filieraagricola.api.entities.elemento.ProdottoTrasformato;
 import it.unicam.cs.filieraagricola.api.services.UserService;
+import it.unicam.cs.filieraagricola.api.services.elemento.ProdottoBaseService;
+import it.unicam.cs.filieraagricola.api.services.elemento.ProdottoService;
 import it.unicam.cs.filieraagricola.api.services.elemento.ProdottoTrasformatoService;
 import it.unicam.cs.filieraagricola.api.services.gestore.CategoriaService;
 import jakarta.annotation.PostConstruct;
@@ -11,6 +15,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import static it.unicam.cs.filieraagricola.api.commons.ResponseEntityUtil.unauthorizedResponse;
+
 @RestController
 @RequestMapping("/operatore/trasformatore")
 public class ProdottoTrasformatoController {
@@ -18,17 +24,14 @@ public class ProdottoTrasformatoController {
     @Autowired
     private ProdottoTrasformatoService prodottoTrasformatoService;
     @Autowired
+    private ProdottoBaseService prodottoBaseService;
+    @Autowired
     private UserService userService;
     @Autowired
     private CategoriaService categoriaService;
 
-    @GetMapping("/dashboard")
-    public ResponseEntity<String> getDashboard() {
-        return ResponseEntity.ok("Benvenuto nella dashboard del Trasformatore");
-    }
-
     @RequestMapping({"/prodottitrasformati"})
-    public ResponseEntity<Object> getProducts() {
+    public ResponseEntity<Object> getProdottiTrasformati() {
         if(!prodottoTrasformatoService.getProdottiTrasformati().isEmpty()){
             return new ResponseEntity<>(prodottoTrasformatoService.getProdottiTrasformati(), HttpStatus.FOUND);
         }
@@ -36,19 +39,27 @@ public class ProdottoTrasformatoController {
     }
 
     @RequestMapping({"/prodottitrasformati/{id}"})
-    public ResponseEntity<Object> getProduct(@PathVariable("id") int id) {
+    public ResponseEntity<Object> getProdottoTrasformato(@PathVariable   int id) {
         if(prodottoTrasformatoService.getProdottoTrasformato(id) != null ){
             if(prodottoTrasformatoService.getProdottoTrasformato(id).getOperatore() == userService.getCurrentUser()) {
                 return new ResponseEntity<>(prodottoTrasformatoService.getProdottoTrasformato(id), HttpStatus.FOUND);
             } else {
-                return new ResponseEntity<>("Non sei autorizzato a vedere questo prodotto", HttpStatus.UNAUTHORIZED);
+                return unauthorizedResponse();
             }
         }
         return new ResponseEntity<>("Prodotto trasformato non esistente", HttpStatus.NOT_FOUND);
     }
 
     @PostMapping({"/prodottitrasformati/aggiungi"})
-    public ResponseEntity<Object> addProduct(@RequestBody ProdottoTrasformato prodotto) {
+    public ResponseEntity<Object> aggiungiProdottoTrasformato(@RequestBody ProdottoTrasformato prodotto) {
+        ProdottoBase prodottoBase = prodottoBaseService.getProdottoBase(prodotto.getProdottoBase().getId());
+        if (prodottoBase == null) {
+            return new ResponseEntity<>("Prodotto base non trovato", HttpStatus.NOT_FOUND);
+        }
+        if (prodottoBase.getStatorichiesta() != StatoContenuto.ACCETTATA) {
+            return new ResponseEntity<>("Il prodotto base associato deve essere stato precedentemente validato", HttpStatus.BAD_REQUEST);
+        }
+
         if(prodottoTrasformatoService.aggiungiProdottoTrasformato(prodotto)){
             return new ResponseEntity<>("Prodotto creato", HttpStatus.CREATED);
         }
@@ -56,7 +67,7 @@ public class ProdottoTrasformatoController {
     }
 
     @PostMapping({"prodottitrasformati/aggiungiconparametri"})
-    public ResponseEntity<Object> addProductWithParam(@RequestParam("nome") String nome,
+    public ResponseEntity<Object> aggiungiProdottoTrasformatoConParametri(@RequestParam("nome") String nome,
                                                       @RequestParam("processoTrasformazione") String processoTrasformazione,
                                                       @RequestParam("certificazioni") String certificazioni,
                                                       @RequestParam("prodottoBase") int IDprodottoBase,
@@ -64,6 +75,13 @@ public class ProdottoTrasformatoController {
                                                       @RequestParam("prezzo") double prezzo,
                                                       @RequestParam("categoria") String categoria)
     {
+        ProdottoBase prodottoBase = prodottoBaseService.getProdottoBase(IDprodottoBase);
+        if (prodottoBase == null) {
+            return new ResponseEntity<>("Prodotto base non trovato", HttpStatus.NOT_FOUND);
+        }
+        if (prodottoBase.getStatorichiesta() != StatoContenuto.ACCETTATA) {
+            return new ResponseEntity<>("Il prodotto base associato deve essere stato precedentemente validato", HttpStatus.BAD_REQUEST);
+        }
         if(!categoriaService.existsSameCategoria(categoria)){
             return new ResponseEntity<>("Categoria non esistente, Categorie esistenti: " + categoriaService.getAllCategorie(), HttpStatus.NOT_FOUND);
         }
@@ -74,25 +92,59 @@ public class ProdottoTrasformatoController {
         return new ResponseEntity<>("Prodotto trasformato esistente", HttpStatus.CONFLICT);
     }
 
-    @RequestMapping({"/prodottitrasformati/elimina/{id}"})
-    public ResponseEntity<Object> deleteProduct(@PathVariable("id") int id) {
-        if(prodottoTrasformatoService.getProdottoTrasformato(id).getOperatore() != userService.getCurrentUser()) {
-            return new ResponseEntity<>("Non sei autorizzato", HttpStatus.UNAUTHORIZED);
-        }
-        prodottoTrasformatoService.deleteProdottoTrasformato(id);
-        return new ResponseEntity<>("Prodotto trasformato eliminato", HttpStatus.OK);
+    @RequestMapping({"/prodottitrasformati/elimina"})
+    public ResponseEntity<Object> eliminaProdottoTrasformato(@RequestParam int id) {
+        if (prodottoTrasformatoService.getProdottoTrasformato(id) != null) {
+            if (prodottoTrasformatoService.getProdottoTrasformato(id).getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            prodottoTrasformatoService.deleteProdottoTrasformato(id);
+            return new ResponseEntity<>("Prodotto trasformato eliminato", HttpStatus.OK);
+        } else return new ResponseEntity<>("Prodotto non esistente", HttpStatus.NOT_FOUND);
     }
 
-    @RequestMapping(
-            value = {"/prodottitrasformati/aggiorna"},
-            method = {RequestMethod.PUT}
-    )
-    public ResponseEntity<Object> updateProduct(@RequestBody ProdottoTrasformato prodottoTrasformato) {
-        if(prodottoTrasformato.getOperatore() != userService.getCurrentUser()) {
-            return new ResponseEntity<>("Non sei autorizzato", HttpStatus.UNAUTHORIZED);
+    @PutMapping("/prodottitrasformati/aggiorna")
+    public ResponseEntity<Object> aggiornaProdottoTrasformato(@RequestBody ProdottoTrasformato prodottoTrasformato) {
+        if (prodottoTrasformatoService.existsProdottoTrasformato(prodottoTrasformato.getId())) {
+            ProdottoTrasformato prodotto = prodottoTrasformatoService.getProdottoTrasformato(prodottoTrasformato.getId());
+            if (prodotto.getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            if (prodottoTrasformatoService.aggiornaProdottoTrasformato(prodottoTrasformato)) {
+                return new ResponseEntity<>("Prodotto trasformato aggiornato", HttpStatus.FOUND);
+            }
+            return new ResponseEntity<>("Prodotto non aggiornato", HttpStatus.NOT_FOUND);
         }
-        if(prodottoTrasformatoService.aggiornaProdottoTrasformato(prodottoTrasformato)){
-            return new ResponseEntity<>("Prodotto trasformato aggiornato", HttpStatus.FOUND);
+        return new ResponseEntity<>("Prodotto non esistente", HttpStatus.NOT_FOUND);
+    }
+
+    @PutMapping("/prodottitrasformati/aggiorna-quantita")
+    public ResponseEntity<Object> aggiornaQuantitaProdotto(@RequestParam int id, @RequestParam int quantita) {
+        ProdottoTrasformato prodotto = prodottoTrasformatoService.getProdottoTrasformato(id);
+        if (prodotto != null) {
+            if (prodotto.getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            if (prodottoTrasformatoService.aggiornaQuantitaProdotto(id, quantita)) {
+                return new ResponseEntity<>("Quantità aggiornata", HttpStatus.OK);
+            }
+            return new ResponseEntity<>("Impossibile aggiornare la quantità", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>("Prodotto non trovato", HttpStatus.NOT_FOUND);
+    }
+
+    @PutMapping("/prodottitrasformati/rimuovi-quantita")
+    public ResponseEntity<Object> rimuoviQuantitaProdotto(@RequestParam int id, @RequestParam int quantita) {
+        ProdottoTrasformato prodotto = prodottoTrasformatoService.getProdottoTrasformato(id);
+        if (prodotto != null) {
+            if (prodotto.getOperatore() != userService.getCurrentUser()) {
+                return unauthorizedResponse();
+            }
+            if (prodottoTrasformatoService.riduciQuantitaProdotto(id, quantita)) {
+                return new ResponseEntity<>("Quantità ridotta", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Quantità insufficiente per rimuovere", HttpStatus.BAD_REQUEST);
+            }
         }
         return new ResponseEntity<>("Prodotto non trovato", HttpStatus.NOT_FOUND);
     }
